server_final/requirements.txt
Django==3.0.14
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.5.0
PyJWT==2.8.0
psycopg2-binary==2.9

deepface
tensorflow-cpu
opencv-python-headless
Pillow
